Change lat/lng type

+++ src/main/java/com/google/sps/data/MapLocation.java
@@ -2,16 +2,16 @@ package com.google.sps.data;
 
 /** Class representing a MapLocation as a coordinate **/
 public class MapLocation {
-  public long lat;
-  public long lng;
+  public double lat;
+  public double lng;
   
   /** Creates a new MapLocation
   * @param lat latitude value
   * @param lng longitude value
   * @return MapLocation object
   **/
-  
-  public MapLocation (long lat, long lng) {
+
+  public MapLocation (double lat, double lng) {
     this.lat = lat;
     this.lng = lng;
   }
